+++ lib/src/_models/postings_map_entry.dart
@@ -4,15 +4,15 @@
 
 import 'package:text_indexing/text_indexing.dart';
 
-/// The [PostingsMapEntry] class enumerates the properties of an entry in a
-/// [PostingsMap] collection:
+/// The [PostingsEntry] class enumerates the properties of an entry in a
+/// [Postings] collection:
 /// - [term] is the word/term that is indexed; and
-/// - [postings] is a hashmap of the [TermPositions] for [term].
-class PostingsMapEntry {
+/// - [postings] is a hashmap of the [PostingsList] for [term].
+class PostingsEntry {
   //
 
-  /// Serializes the [PostingsMapEntry] to a MapEntry for direct insertion in
-  /// a [PostingsMap].
+  /// Serializes the [PostingsEntry] to a MapEntry for direct insertion in
+  /// a [Postings].
   MapEntry<String, Map<String, List<int>>> toMapEntry() => MapEntry(
       term,
       postings
@@ -21,24 +21,24 @@ class PostingsMapEntry {
   /// The word/term that is indexed.
   final String term;
 
-  /// A hashmap of the [TermPositions] for the [term] where:
+  /// A hashmap of the [PostingsList] for the [term] where:
   /// - the [postings].key is the id of the document;
   /// - the [postings].value is the positions in the document of the [term].
-  final Map<String, TermPositions> postings;
+  final Map<String, PostingsList> postings;
 
-  /// Instantiates a const [PostingsMapEntry] instance:
+  /// Instantiates a const [PostingsEntry] instance:
   /// - [term] is the word/term that is indexed; and
-  /// - [postings] is a hashmap of the [TermPositions] for [term].
-  const PostingsMapEntry(this.term, this.postings);
+  /// - [postings] is a hashmap of the [PostingsList] for [term].
+  const PostingsEntry(this.term, this.postings);
 
-  /// Factory constructor that instantiates a [PostingsMapEntry] instance from
-  /// a [PostingsMap] map entry.
-  factory PostingsMapEntry.fromEntry(
+  /// Factory constructor that instantiates a [PostingsEntry] instance from
+  /// a [Postings] map entry.
+  factory PostingsEntry.fromEntry(
       MapEntry<String, Map<String, List<int>>> entry) {
     final term = entry.key;
     final postings = entry.value.map((key, value) =>
-        MapEntry(key, TermPositions.fromEntry(term, MapEntry(key, value))));
+        MapEntry(key, PostingsList.fromEntry(term, MapEntry(key, value))));
 
-    return PostingsMapEntry(term, postings);
+    return PostingsEntry(term, postings);
   }
 }