+++ lib/text_indexing.dart
@@ -10,5 +10,4 @@ export 'src/_extensions/_index.dart';
 export 'src/_models/_index.dart';
 export 'src/persisted_indexer.dart';
 export 'src/in_memory_indexer.dart';
-export 'src/indexer.dart';
-export 'src/typedefs.dart';
+export 'src/text_indexer.dart';