+++ CHANGELOG.md
@@ -4,6 +4,21 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+## 0.0.1
+
+**PRE-RELEASE, BREAKING CHANGES**
+- Interfaces finalized
+- *BREAKING CHANGES:* `TextIndexerBase`.
+- *BREAKING CHANGES:* `InMemoryIndexer`.
+- *BREAKING CHANGE:* `TermDictionary` renamed `Dictionary`.
+- *BREAKING CHANGE:* `PostingsMap` renamed `Postings`.
+- *BREAKING CHANGE:* `PostingsMapEntry` renamed `PostingsEntry`.
+- *BREAKING CHANGE:* `Term` renamed `DictionaryEntry`.
+- *BREAKING CHANGE:* `TermPositions` renamed `PostingsList`.
+- Added a test for [PersistedIndexer].
+- Added an example for [PersistedIndexer].
+- Updated documentation.
+
 ## 0.0.1-beta.4
 
 - **PRE-RELEASE**
@@ -12,9 +27,9 @@ All rights reserved.
 ## 0.0.1-beta.3
 
 - **PRE-RELEASE, BREAKING CHANGES**
-- *BREAKING CHANGES:* `Indexer`.
+- *BREAKING CHANGES:* `TextIndexerBase`.
 - *BREAKING CHANGES:* `InMemoryIndexer`.
-- Coded `PersistedIndexer`.
+- Added `PersistedIndexer`.
 - Updated examples.
 - Updated tests.
 - Updated documentation.
@@ -22,8 +37,8 @@ All rights reserved.
 ## 0.0.1-beta.2
 
 - **PRE-RELEASE**
-- Coded `Indexer`.
-- Coded `InMemoryIndexer`.
+- Added `TextIndexerBase`.
+- Added `InMemoryIndexer`.
 - Created test of `InMemoryIndexer`.
 - Created example of `InMemoryIndexer`.
 - Updated documentation.